Added animated loader
