+++ sites/partners/pages/application/edit.tsx
@@ -1,15 +1,7 @@
 import React from "react"
 import Head from "next/head"
 import { useRouter } from "next/router"
-import {
-  AppearanceStyleType,
-  Button,
-  MetaTags,
-  PageHeader,
-  StatusBar,
-  t,
-} from "@bloom-housing/ui-components"
-import { ApplicationStatus } from "@bloom-housing/backend-core/types"
+import { MetaTags, PageHeader, t } from "@bloom-housing/ui-components"
 import Layout from "../../layouts/application"
 import PaperApplicationForm from "../../src/applications/PaperApplicationForm/PaperApplicationForm"
 import { useSingleApplicationData } from "../../lib/hooks"
@@ -47,27 +39,6 @@ const NewApplication = () => {
           }
         />
 
-        <StatusBar
-          backButton={
-            <Button
-              inlineIcon="left"
-              icon="arrow-back"
-              onClick={() => router.push(`/application?id=${application.id}`)}
-            >
-              {t("t.back")}
-            </Button>
-          }
-          tagStyle={
-            application?.status == ApplicationStatus.submitted
-              ? AppearanceStyleType.success
-              : AppearanceStyleType.primary
-          }
-          tagLabel={
-            application?.status
-              ? t(`application.details.applicationStatus.${application.status}`)
-              : t(`application.details.applicationStatus.submitted`)
-          }
-        />
         <PaperApplicationForm
           listingId={application.listing.id}
           application={application}
